[data] make some _aggregate api internal
