+++ doc/source/data/api/grouped_data.rst
@@ -49,9 +49,3 @@ Aggregate Function
    :toctree: doc/
 
    aggregate.AggregateFn
-   aggregate.Count
-   aggregate.Sum
-   aggregate.Max
-   aggregate.Mean
-   aggregate.Std
-   aggregate.AbsMax
