+++ release/nightly_tests/dataset/aggregate_benchmark.py
@@ -1,8 +1,13 @@
 from typing import Tuple
 
 import ray
-from ray.data.aggregate import Max, Mean, Min, Sum
-from ray.data.aggregate._aggregate import _AggregateOnKeyBase
+from ray.data._internal.aggregate._aggregate import (
+    _AggregateOnKeyBase,
+    Max,
+    Mean,
+    Min,
+    Sum,
+)
 from ray.data.block import Block
 from ray.data.dataset import Dataset
 import pyarrow.compute as pac
